refactor(tps): 更新 tps 模块路径
- 将 `tpsShower.ts` 从 `plugins/` 移动到 `core/tps` 目录
- 更新 `main.ts` 中的模块导入路径
